Move WhelkTool to correct directory for package declaration
