+++ housekeeping/src/main/groovy/whelk/housekeeping/ScriptRunner.java
@@ -1,17 +1,14 @@
 package whelk.housekeeping;
 
-import datatool.WhelkTool;
 import org.apache.commons.io.IOUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import whelk.Whelk;
-import whelk.housekeeping.HouseKeeper;
+import whelk.datatool.WhelkTool;
 
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
