+++ whelktool/src/test/groovy/whelk/datatool/util/DisjointSetsTest.java
@@ -1,4 +1,4 @@
-package datatool.util;
+package whelk.datatool.util;
 
 import static org.junit.Assert.assertEquals;
