+++ whelktool/src/test/groovy/whelk/datatool/util/DocumentComparatorSpec.groovy
@@ -1,4 +1,4 @@
-package datatool.util
+package whelk.datatool.util
 
 
 import spock.lang.Specification
